+++ packages/library-sync/src/commonMain/kotlin/io/realm/kotlin/mongodb/internal/AppConfigurationImpl.kt
@@ -48,12 +48,14 @@ public class AppConfigurationImpl @OptIn(ExperimentalKBsonSerializerApi::class) 
     internal val networkTransportFactory: (dispatcher: DispatcherHolder) -> NetworkTransport,
     override val metadataMode: MetadataMode,
     override val syncRootDirectory: String,
-    public val logger: LogConfiguration,
+    public val logger: LogConfiguration?,
     override val appName: String?,
     override val appVersion: String?,
     internal val bundleId: String,
     override val ejson: EJson,
-    override val httpLogObfuscator: HttpLogObfuscator?
+    override val httpLogObfuscator: HttpLogObfuscator?,
+    override val customRequestHeaders: Map<String, String>,
+    override val authorizationHeaderName: String,
 ) : AppConfiguration {
 
     /**